requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dj-config-url==0.1.1
Django==4.2.5
django-environ==0.11.2
django-extensions==3.2.1
idna==3.4
kombu==5.3.2
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
transliterate==1.10.2
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
